Code review with important additions to the ignore file for python - python is a curious beast
