+++ .gitignore
@@ -2,3 +2,6 @@
 ./node_modules
 node_modules
 client/instalist/node_modules/
+__pycache__/
+config/__pycache__/
+controllers/__pycache__/
